+++ src/modules/dayidsub.ts
@@ -1,7 +1,7 @@
 import axios, { Axios } from 'axios';
 import * as jsdom from 'jsdom';
 import { AnimeSubsApiResponse } from '../interfaces';
-import { AxiosClient } from './axiosClient';
+import { AxiosClient } from '../api/axiosClient';
 
 const { JSDOM } = jsdom;
 const virtualConsole = new jsdom.VirtualConsole();
