+++ src/modules/okamisubs.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import * as jsdom from 'jsdom';
-import { AxiosClient } from './axiosClient';
+import { AxiosClient } from '../api/axiosClient';
 
 const { JSDOM } = jsdom;
 const virtualConsole = new jsdom.VirtualConsole();
